allows unittest lib folder
This commit allows the lib folder in unittest which was previously not
included due to .gitignore.   This commit brings the common test lib
functions and updates the system tests to inheret from DutSystemTest.

The DutSystemTest base uses dut.conf in test/fixtures to determine which
nodes to run the system tests against.