refactored switchports module to sources
This commit refactors the old switchports module to use the client resource
calls.   Also adds system and unit tests for the switchport module.  Below
is an example of how to use the switchport resource

>>> import pyeapi.client
>>> pyeapi.client.load_config()
>>> node = pyeapi.connect_to('veos01')
>>> ports = node.resource('switchports')
>>> ports.getall()

The switchports resource allows for the configuration of logical layer 2
interfaces on Ethernet and Port-Channel physical interfaces