adds support for unix domain sockets and http local connections
This commit introduces a new eapilib for handling all variants of eAPI
transports.   It adds support for connecting via Unix domain sockets and
HTTP local connections introduced in 4.14.5F.

This to use the new connection types, the conf file has been modifed to
add a new recognized parameter: connection_type.  The connection_type
parameter takes one of unix, http_local, http, https.   If the
connection_type parameter is not specified, the it defaults to http.