fixes issue with not properly loading conf file
This commit addresses an issue that would prevent the client from loading
the conf file from the users home folder.  The load_config method did not
properly expand the ~ for obtaining the full path to the conf file.

This commit is required to support loading conf files from the users
home folder