refactors the base abstract classes out of the package init
This commit refactors the base abstract classes out of the package init
module and into pyeapi.api.abstract.  The classes in abstract should not be
directly instantiated.  The clases should be used as base classes for building
API modules that work with the Node object.

The abstract classes have been imported into the pyeapi.api package for
streamlined operation and to provide backwards compabilility