adds new functionality to the Node enable method
This commit redefines how the enable method for a Node instance works.  The
enable method will now automatically try to run a command with both json and
text encoding (assuming the former wasn't converted).   This will add
a new keyword arg (strict) to enable.

To use the command in the strict form, the method will not try the to
run the command with text encoding, rather will allow the CommandError
to be raised.