refactors config into the node object
This commit refactors the config out of the api and into node to make it
globally accessible via the node obect.  This also changes the default
operation to enable autorefresh on config changes.  The autorefresh can
still be diabled by setting autorefresh to False in the node instance.

The configuration is now shared across all api instances so if one api call
updates the configuration all entities will see the change (assuming
autorefresh has not be disabled)

This commit also renamed the user_authentication method in the connection
object to just authentication and the exec_authentication method is now
enable_authentication.