adds exclude_interfaces setting to conf file for testing
This commit adds the ability to specify interfaces to be excluded from
being choosen by a call to random_interface.  The setting can be configured
in the conf file by specifying a list of comma separated interfaces names.

For example, to exclude ethernet1-3 make the following entry in the
eapi.conf file:

[connection:localhost]
exclude_interfaces: Ethernet1, Ethernet2, Ethernet3

Note the full interface name must be specified and only Ethernet interfaces
are considered