implements a new config object for managing the pyeapi conf file
This commit creates a new Config object in pyeapi.client that will manage
the conf file.  The implementation should be comletely transparent to the
API and you can still use load_config and config_for. There is one new
feature that was added.  The client will now automatically load the conf
file when it is first imported.   This eliminates the step to have to make
a call to load_config.  The client can still use load_config without any
adverse effects and can still pass a filepath to load_config to load
a specific file.