adds additional exception handling for http connections
This commit adds an additional exception handling for decoding the JSON
response from the node.  In the case addressed, a valid HTTP response is
returned from the node but the payload raises a ValueError when trying to
decode it.  This patch will now return a ConnectionError in this case
and provide the response message in the exception.

try:
    node.enable('show version')
except ConnectionError as exc:
    print exc.response

'Unable to authentiate user: Bad username/password combination'