adds support for set_trunk_groups to the vlans api module
This commit adds a new method to the vlans api module set_trunk_groups.  The
method will take as list of trunk group names and add and/or remove trunk
groups as needed in the config.  It also accepts no value to negate the
trunk groups and accepts the default flag to default the config.

set_trunk_groups('10', ['foo', 'bar'])
set_trunk_groups('10')
set_trunk_groups('10', default=True)

This commit only adds the new method and associated test cases and should be
generally safe to apply.