fixes two issues with ipinterfaces
This commit address two minor issues with the ipinterfaces api module

The first issue is a problem where the get method would not be able to
find an ip address in the interface config.  This occurs if the interface
has been configured but an ip address is not yet set.  The fix will now
return 0.0.0.0 for the address attribute in this case

The second issue is an unexpected ValueError exception with configing the
MTU setting.  The set_mtu method now corectly validates the MTU range and
will not trigger the ValueError with a valid MTU setting.

This commit includes the above fixes and updatees to the system test cases
to check for these issues.