adds introspection into CommandError exception
When a CommandError is raised, there is additional information inside the
response messages that is now captured in the CommandError exception
instance.  This change will capture the inner error as well as provide
a command trace that shows the output from the commands leading up to the
error message.  The trace will also show the commands that didn't run which
are identified by setting the output to None as opposed to an empty
dict object.

This commit adds two new public properties to the CommandError exception
instance:

* command_error - the inner error text message
* trace - a list of dictionaries showing the command executed and the ouput
from the command

This commit was added as an enhacement based on issue #4