updates the config method to make a copy of the commands list
This commit updates the config method to make a copy of the commands list
begin sent to the transport.  Without this commit, the commands list object
was updated directly with enable and configure which could cause issues
in the implementation.  Making a copy of the commands list first will fix
this issue.