adds support for py2.7.9 default certification verification
Python 2.7.9 added certification verification by default as a backport from
Python 3.  This changes the default behavior of SSL certification verification.
This patch will restore the former behavior to not fail on ceritifcation
validation for self signed certificates.

This patch does not enfore verification by default on Python 2.7.9 or later
in order to not break compatability.  This default behavior will be changed
in later releases of pyeapi.

For Python version prior to 2.7.9, nothing changes